pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "faiss-cpu>=1.9.0.post1",
    "numpy>=2.2.1",
    "flask-wtf>=1.2.2",
    "werkzeug>=3.1.3",
    "oauthlib>=3.2.2",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "scikit-learn>=1.6.0",
    "twilio>=9.4.1",
    "aiogram==3.3.0",
]
